feat: Add getSlot and getEvaluatees functions
This commit adds two new functions, `getSlot` and `getEvaluatees`, to the `page.tsx` file. These functions are responsible for retrieving the evaluation slot and evaluatees data from the database. The `getSlot` function takes in the `id` and `evaluatorUserId` parameters and returns the corresponding evaluation slot. The `getEvaluatees` function takes in the `evaluationSlotId` parameter and returns an array of evaluatees associated with that evaluation slot. This code change improves the modularity and readability of the code.